+++ metrics_utility/automation_controller_billing/collectors.py
@@ -14,7 +14,7 @@ from django.utils.timezone import now, timedelta
 from django.utils.translation import gettext_lazy as _
 
 from metrics_utility.base import CsvFileSplitter, register
-from metrics_utility.base.collector import get_max_gather_period_days
+from metrics_utility.base.utils import get_max_gather_period_days
 from metrics_utility.logger import logger
